requirements.txt
numpy==1.21.3
pandas==1.3.4
rdkit-pypi==2021.9.2
scikit-learn==0.20.3
scipy==1.2.1
tensorboard==1.13.1
tensorflow==1.13.1
protobuf=3.20.2
